+++ scala/ppml/src/main/scala/com/intel/analytics/bigdl/ppml/FLServer.scala
@@ -21,7 +21,8 @@ import com.intel.analytics.bigdl.ppml.common.Aggregator
 import com.intel.analytics.bigdl.ppml.psi.PSIServiceImpl
 import java.io.{File, IOException}
 
-import com.intel.analytics.bigdl.ppml.service.NNServiceImpl
+import com.intel.analytics.bigdl.ppml.fgboost.FGBoostServiceImpl
+import com.intel.analytics.bigdl.ppml.nn.NNServiceImpl
 import org.apache.logging.log4j.{Level, LogManager}
 import org.apache.logging.log4j.core.config.Configurator
 
@@ -66,5 +67,6 @@ class FLServer private[ppml](val _args: Array[String] = null) extends GrpcServer
     // start all services without providing service list
     serverServices.add(new PSIServiceImpl(clientNum))
     serverServices.add(new NNServiceImpl(clientNum))
+    serverServices.add(new FGBoostServiceImpl(clientNum))
   }
 }
